requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
pyee==7.0.4
requests==2.24.0
slackclient==2.9.2
slackeventsapi==2.2.1
soupsieve==2.0.1
urllib3==1.25.11
Werkzeug==1.0.1
yarl==1.5.1
